requirements.txt
streamlit
python-docx
crewai
langchain_openai
langchain
langchain_community